Centralise a valid config for testing This commit introduces tests/conftest.py, the main purpose of which is to centralise a valid config dict that is shared by all core and gui test functions.
