+++ tests/gui/conftest.py
@@ -1,75 +1,18 @@
-"""Shared fixtures for GUI tests."""
+"""Shared fixtures for GUI tests.
+
+`valid_config` and `working_dir` come from the top-level tests/conftest.py -
+pytest injects them automatically, no import needed.
+"""
 
 # pylint doesn't recognise pytest's fixture injection
 # pylint: disable=redefined-outer-name
 
 from copy import deepcopy
-from pathlib import Path
 
 import pytest
 
 from openairclim.gui import config_io
 from openairclim.gui.state import AppState
-
-CORE_TESTS_DIR = Path(__file__).resolve().parents[1] / "core"
-REPO_DIR = CORE_TESTS_DIR / "repository"
-BG_NAME = "co2_bg.nc"
-INV_NAME = "test_inv.nc"
-
-
-def make_valid_config():
-    """Return a complete, valid configuration dict.
-
-    Mirrors `read_config_test.TestCheckConfig.test_correct_config`: every path
-    is relative to `CORE_TESTS_DIR`, which is used as the working directory
-    wherever this config is exercised.
-
-    Returns:
-        dict: A configuration dict that passes `core.read_config.check_config`.
-    """
-    repo_path = "repository/"
-    return {
-        "species": {"inv": ["CO2"], "nox": "NO", "out": ["CO2"]},
-        "inventories": {
-            "dir": repo_path,
-            "files": [INV_NAME],
-            "rel_to_base": False,
-            "base": {"dir": repo_path, "files": [INV_NAME]},
-        },
-        "output": {
-            "run_oac": True,
-            "run_metrics": True,
-            "run_plots": True,
-            "dir": "results/",
-            "name": "example",
-            "overwrite": True,
-            "concentrations": False,
-        },
-        "time": {"range": [2020, 2121, 1]},
-        "background": {
-            "dir": repo_path,
-            "CO2": {"file": (repo_path + BG_NAME), "scenario": "SSP2-4.5"},
-            "CH4": {"file": (repo_path + BG_NAME), "scenario": "SSP2-4.5"},
-            "N2O": {"file": (repo_path + BG_NAME), "scenario": "SSP2-4.5"},
-        },
-        "responses": {"dir": repo_path},
-        "temperature": {"method": "Boucher&Reddy", "CO2": {"lambda": 1.0}},
-        "metrics": {"types": ["ATR"], "t_0": [2020], "H": [100]},
-        "aircraft": {"types": ["DEFAULT"]},
-    }
-
-
-@pytest.fixture
-def valid_config():
-    """A complete, valid configuration dict."""
-    return make_valid_config()
-
-
-@pytest.fixture
-def working_dir():
-    """Working directory that `valid_config`'s relative paths resolve
-    against."""
-    return str(CORE_TESTS_DIR)
 
 
 @pytest.fixture
